renamed message.py file to email_intent.py
