+++ src/email/domain/models/email_intent.py
